tests: Restructure sysinfo tests
