+++ tests/meson.build
@@ -14,8 +14,8 @@ nvapi_src = files([
 
 nvapi_tests_src = files([
   'main.cpp',
-  'nvapi.cpp',
   'nvapi_d3d11.cpp',
+  'nvapi_sysinfo.cpp',
 ])
 
 nvapi_exe = executable('nvapi'+target_suffix+'-tests', [ nvapi_src, nvapi_tests_src, dxvk_nvapi_version ],
